+++ tests/system/example.spec.ts
@@ -47,7 +47,7 @@ describe("Example", async () => {
       expect(response.headers.has("access-control-allow-headers")).toBeTruthy();
       expect(response.headers.get("access-control-allow-origin")).toBe("*");
       expect(response.headers.get("access-control-allow-methods")).toBe(
-        "POST, OPTIONS",
+        "PATCH, OPTIONS",
       );
       expect(response.headers.get("access-control-allow-headers")).toBe(
         "content-type",
@@ -55,8 +55,22 @@ describe("Example", async () => {
     });
 
     test("Should handle valid POST request", async () => {
-      const response = await fetch(`http://localhost:${port}/v1/user/50`, {
+      const response = await fetch(`http://localhost:${port}/v1/user/create`, {
         method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ name: "John Doe" }),
+      });
+      expect(response.status).toBe(201);
+      const json = await response.json();
+      expect(json).toMatchObject({
+        status: "created",
+        data: { id: 16 },
+      });
+    });
+
+    test("Should handle valid PATCH request", async () => {
+      const response = await fetch(`http://localhost:${port}/v1/user/50`, {
+        method: "PATCH",
         headers: {
           token: "456",
           "Content-Type": "application/json",
@@ -76,7 +90,7 @@ describe("Example", async () => {
           createdAt: "2022-01-22T00:00:00.000Z",
         },
       });
-      await waitFor(() => /v1\/user/.test(out));
+      await waitFor(() => /v1\/user\/:id/.test(out));
       await waitFor(() => /50, 123, 456/.test(out));
       expect(true).toBeTruthy();
     });
@@ -243,9 +257,31 @@ describe("Example", async () => {
       expect(true).toBeTruthy();
     });
 
-    test("POST request should fail on auth middleware key check", async () => {
-      const response = await fetch(`http://localhost:${port}/v1/user/50`, {
+    test("POST request should respond with a conflict on assertion of uniqueness", async () => {
+      const response = await fetch(`http://localhost:${port}/v1/user/create`, {
         method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ name: "James McGill" }),
+      });
+      expect(response.status).toBe(409);
+      const json = await response.json();
+      expect(json).toEqual({ status: "exists", id: 16 });
+    });
+
+    test("POST request should fail on demand", async () => {
+      const response = await fetch(`http://localhost:${port}/v1/user/create`, {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ name: "Gimme Jimmy" }),
+      });
+      expect(response.status).toBe(500);
+      const json = await response.json();
+      expect(json).toEqual({ status: "error", reason: "That went wrong" });
+    });
+
+    test("PATCH request should fail on auth middleware key check", async () => {
+      const response = await fetch(`http://localhost:${port}/v1/user/50`, {
+        method: "PATCH",
         headers: {
           token: "456",
           "Content-Type": "application/json",
@@ -266,9 +302,9 @@ describe("Example", async () => {
       });
     });
 
-    test("POST request should fail on auth middleware token check", async () => {
+    test("PATCH request should fail on auth middleware token check", async () => {
       const response = await fetch(`http://localhost:${port}/v1/user/50`, {
-        method: "POST",
+        method: "PATCH",
         headers: {
           token: "123",
           "Content-Type": "application/json",
@@ -289,9 +325,9 @@ describe("Example", async () => {
       });
     });
 
-    test("POST request should fail on schema validation", async () => {
+    test("PATCH request should fail on schema validation", async () => {
       const response = await fetch(`http://localhost:${port}/v1/user/-50`, {
-        method: "POST",
+        method: "PATCH",
         headers: {
           token: "456",
           "Content-Type": "application/json",
@@ -307,9 +343,9 @@ describe("Example", async () => {
       expect(json).toMatchSnapshot();
     });
 
-    test("POST request should fail on specific value in handler implementation", async () => {
+    test("PATCH request should fail on specific value in handler implementation", async () => {
       const response = await fetch(`http://localhost:${port}/v1/user/101`, {
-        method: "POST",
+        method: "PATCH",
         headers: {
           token: "456",
           "Content-Type": "application/json",